move discordmanager to nebulaworld
maybe it's more appropriate here

if anything it allows for the hacky method of determining the number of players by looking at remote player models
