+++ NebulaNetwork/NebulaNetwork.csproj
@@ -4,7 +4,6 @@
     <Compile Remove="PacketProcessors\Session\LobbyUpdateValues.cs" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\dep\discord_game_sdk\discord_game_sdk_dotnet.csproj" />
     <ProjectReference Include="..\NebulaModel\NebulaModel.csproj" SpecificVersion="true" />
     <ProjectReference Include="..\NebulaAPI\NebulaAPI.csproj" Private="false" />
     <ProjectReference Include="..\NebulaWorld\NebulaWorld.csproj" SpecificVersion="true" />
